Updated: 구조 개선 [Exception 계층 분리]
