--- src/main/kotlin/com/hhplus/ecommerce/common/exception/balance/BalanceNotFoundException.kt
@@ -1,3 +1,0 @@
-package com.hhplus.ecommerce.common.exception.balance
-
-class BalanceNotFoundException(message: String = "잔액 정보가 존재하지 않습니다."): BalanceException(message)
