--- src/main/kotlin/com/hhplus/ecommerce/common/exception/balance/InsufficientBalanceException.kt
@@ -1,3 +1,0 @@
-package com.hhplus.ecommerce.common.exception.balance
-
-class InsufficientBalanceException(message: String = "잔액이 부족합니다."): BalanceException(message)
