--- src/main/kotlin/com/hhplus/ecommerce/common/exception/payment/ExternalPaymentException.kt
@@ -1,3 +1,0 @@
-package com.hhplus.ecommerce.common.exception.payment
-
-class ExternalPaymentException(message: String = "외부 PG사 결제에서 오류가 발생했습니다."): PaymentException(message)
