--- src/main/kotlin/com/hhplus/ecommerce/common/exception/product/OutOfStockException.kt
@@ -1,3 +1,0 @@
-package com.hhplus.ecommerce.common.exception.product
-
-class OutOfStockException(message: String = "상품 재고가 부족합니다."): ProductException(message)
