+++ src/main/kotlin/com/hhplus/ecommerce/infrastructure/balance/BalanceRepository.kt
@@ -1,6 +1,6 @@
 package com.hhplus.ecommerce.infrastructure.balance
 
-import com.hhplus.ecommerce.common.exception.balance.BalanceNotFoundException
+import com.hhplus.ecommerce.infrastructure.balance.exception.BalanceNotFoundException
 import com.hhplus.ecommerce.domain.balance.respository.IBalanceRepository
 import com.hhplus.ecommerce.infrastructure.balance.jpa.BalanceJpaRepository
 import com.hhplus.ecommerce.infrastructure.balance.jpa.entity.BalanceEntity
