+++ src/main/kotlin/com/hhplus/ecommerce/infrastructure/order/OrderRepository.kt
@@ -1,6 +1,6 @@
 package com.hhplus.ecommerce.infrastructure.order
 
-import com.hhplus.ecommerce.common.exception.order.OrderNotFoundException
+import com.hhplus.ecommerce.infrastructure.order.exception.OrderNotFoundException
 import com.hhplus.ecommerce.domain.order.repository.IOrderRepository
 import com.hhplus.ecommerce.infrastructure.order.jpa.OrderJpaRepository
 import com.hhplus.ecommerce.infrastructure.order.jpa.entity.OrderEntity
