+++ src/test/kotlin/com/hhplus/ecommerce/infrastructure/balance/jpa/entity/BalanceEntityTest.kt
@@ -1,7 +1,7 @@
 package com.hhplus.ecommerce.infrastructure.balance.jpa.entity
 
-import com.hhplus.ecommerce.common.exception.balance.BalanceLimitExceededException
-import com.hhplus.ecommerce.common.exception.balance.InsufficientBalanceException
+import com.hhplus.ecommerce.infrastructure.balance.exception.BalanceLimitExceededException
+import com.hhplus.ecommerce.infrastructure.balance.exception.InsufficientBalanceException
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
